feat: 체크리스트 섹션 지역 에러 바운더리 추가
